refactor(proxy): extract request handler and Clerk instances into lib/auth modules
Behavior-preserving decomposition of the 909-line proxy.ts middleware:
- lib/auth/proxy-request-handler.ts: handleRequest (now handleProxyRequest)
  plus its private helpers (bot detection, CSP nonce, redirect circuit breaker)
- lib/auth/clerk-middleware-instances.ts: production/staging Clerk middleware
  instances, mock-config detection, selectClerkMiddleware()
- auth-degraded-fallback.ts: respondAuthDegraded() collapses the 3x repeated
  HTML/JSON degraded-response block
- circuit-breaker fallthrough (captureWarning + cookie cleanup) deduped into
  applyStateRedirectOrBreak()

No logic, ordering, status-code, or matcher changes. All 619 middleware and
lib/auth tests pass unchanged.
